Update go_bulk_stager.py
This patch removes the code for special-casing the Rucio scope of log files. 

On Perlmutter, we observed issues where log files were often put into the wrong rucio scope. After commenting out this code, the issues generally went away. 

The intent of the original code is not clear to me. 
If necessary I can keep an out-of-tree patch for Perlmutter.